pyproject.toml
[tool.poetry]
name = "grievios"
version = "0.1.0"
description = ""
authors = ["Hannes Hauer <grievios@hanneshauer.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.15"
appium-python-client = "^2.10.1"
networkx = "^3.1"

[tool.poetry.scripts]
grievios = "grievios.main:start"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
